Feature/multi net support (#670)
* multi autoclaim support + api

* Check reorg fix

* scripts + mt extra test

* e2e test

* AutoClaim Compressor working with multi rollup

* refactor + unit tests + compressor e2e using multi-single-bridge

* linter

* fix e2e

* fix unit tests

* Fix db

* fix tests

* fix e2e

* fix benchmark test

* remove TODO

* comment

+++ claimtxman/types/monitoredtx.go
@@ -51,7 +51,7 @@ func (s MonitoredTxStatus) String() string {
 // plus information to monitor if the transactions was sent successfully
 type MonitoredTx struct {
 	// DepositID is the tx identifier controller by the caller
-	DepositID uint
+	DepositID uint64
 
 	// From is a sender of the tx, used to identify which private key should be used to sing the tx
 	From common.Address
@@ -171,7 +171,7 @@ func NewMonitoredTxGroup(entry MonitoredTxGroupDBEntry, txs []MonitoredTx) Monit
 	return res
 }
 
-func (m MonitoredTxGroup) GetTxByDepositID(depositID uint) *MonitoredTx {
+func (m MonitoredTxGroup) GetTxByDepositID(depositID uint64) *MonitoredTx {
 	for idx := range m.Txs {
 		if m.Txs[idx].DepositID == depositID {
 			return &m.Txs[idx]
